store/buyer/requirements.txt
PyMySQL==1.0.2
cryptography==37.0.4
Flask==2.2.2
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.4.4